ref: add setFilters callback
