+++ app/javascript/react/components/SessionFilters/SessionTypeToggle.tsx
@@ -7,54 +7,20 @@ import { useAppDispatch } from "../../store/hooks";
 import { setLoading } from "../../store/mapSlice";
 import { resetUserThresholds } from "../../store/thresholdSlice";
 import { SessionType, SessionTypes } from "../../types/filters";
-import { UserSettings } from "../../types/userStates";
 import { UrlParamsTypes, useMapParams } from "../../utils/mapParamsHandler";
-import useMobileDetection from "../../utils/useScreenSizeDetection";
 import { FilterInfoPopup } from "./FilterInfoPopup";
 import * as S from "./SessionFilters.style";
 
 const SessionTypeToggle = () => {
   const dispatch = useAppDispatch();
-  const { currentUserSettings, searchParams, sessionType, setUrlParams } =
-    useMapParams();
-  const isMobile: boolean = useMobileDetection();
+  const { searchParams, sessionType, setFilters } = useMapParams();
   const { t } = useTranslation();
 
   const handleClick = useCallback(
     (type: SessionType) => {
       dispatch(resetUserThresholds());
       dispatch(setLoading(true));
-      if (isMobile) {
-        setUrlParams([
-          {
-            key: UrlParamsTypes.sessionType,
-            value: type,
-          },
-        ]);
-      } else {
-        setUrlParams([
-          {
-            key: UrlParamsTypes.sessionType,
-            value: type,
-          },
-          {
-            key: UrlParamsTypes.previousUserSettings,
-            value: currentUserSettings,
-          },
-          {
-            key: UrlParamsTypes.currentUserSettings,
-            value: UserSettings.MapView,
-          },
-          {
-            key: UrlParamsTypes.sessionId,
-            value: "",
-          },
-          {
-            key: UrlParamsTypes.streamId,
-            value: "",
-          },
-        ]);
-      }
+      setFilters(UrlParamsTypes.sessionType, type);
     },
     [searchParams]
   );
